removed old DB versions
